[Mojo][Examples] Port examples and description from Programming Manual to HelloMojo.ipynb. (#12999)
This notebook eventually will become a "Mojo Guided Tour" - a page users
would start from.

Currently the examples are copied almost verbatim from the Programming
Manual, and do not always work. Eventually we should try to make them
runnable in the notebook setting.

The difference between this notebook and Programming Manual is that the
notebook does not include rationale and tries to provide a terse
overview of Mojo features, while the Programming Manual offers a
detailed description and motivation for why each feature exists.

modular-orig-commit: 185ef377522942debaa35ed3f660335406e3f70a