[mojo-lang] Add lazy-init for 'let' declarations, warn about unmutated 'var's (#13127)
This enhances check-lifetimes to keep track of whether a declared value
is
mutated after initialization. If so and the value was declared as a
'let'
then this is an error that is diagnosed with a nice error message. If
not
and the value was declared as a 'var', a warning is produced to nudge it
to being declared as a 'let'.

As part of this, we now support lazy initializations of lets, e.g.:

```
     let x : Int
     if cond:
       x = foo()
     else:
       x = bar()
     use(x)
```

As part of this, we mark a bunch more stuff as let's.

We don't currently support proper tracking of uses in closures, so the
warning is disabled for functions containing nested functions.  We need
to improve our closure model in general.

modular-orig-commit: 090ea379003d44ff113035b105c9be55d987b23c