[mojo-lang] Reimplement support for `_` as a DLValue. (#13142)
We previously implement _ as a synthesized vardecl node.  This
changes it to be a dynamic lvalue whose setter just converts the
RHS to an RValue and discards it.  This avoids errors when discarding
values that don't have a move/copy ctor, and generates less IR.

This patch also merges together a few expr nodes for literals since
they are so similar, which is just a refactoring.

This fixes #13073

modular-orig-commit: 8411457b2560c6287eb02d159c3f73517eb6408c