[mojo-lang] Change auto dtors to `__del__` instead of `__del___`. (#13186)
This changes the compiler over to activating on the method of the
right name.  Existing code still has a number of the old pattern
which we should erode over time, so I renamed them all to `del_old()`
for clarity.  Several of them (e.g. Runtime) are more context-manager'y
than auto-dtor'y, so will need new language features to replace.

modular-orig-commit: 93f8a4fc94da1ad9777ba58d1d5549385a9e4a69