[Mojo][Docs] Copyedit Hello Mojo notebook (pt 1) (#13408)
While working through this, I realized that if I want to copyedit stuff
that's reused from the programming manual, I should focus on the origin
document first. So I'll commit more updates to this notebook in chunks.
This is part 1, dealing with the intro and most of the systems section.

modular-orig-commit: 320a69eb016d4ab18d1924d902ec1829cf3ced14