[mojo][docs] Add a build target MojoPublicNotebookGen for processing notebooks (#13409)
To enable proper execution testing, we've been writing `CHECK`s and
other metadata inside of our public notebooks. We don't actually want
to show this when rendering or exporting the notebooks, so we need to
strip it out. This PR adds a build target that processes the notebooks
for export, outputing them to the build directory.

Fixes #13326

modular-orig-commit: d71fd6fa3d88162bc68d018da555e0419632ae29