[mojo-lang][docs] Make the HelloMojo autotune example actually search (#15515)
This section talked about autotuning, user-defined evaluators, and
search. However, the search does not actually get performed, and the
top-level call to `buffer_elementwise_add` was just getting replaced
with one of the implementations with no search.

This PR adds in an actual evaluator, showing users how those things get
defined and that it's not just "magical compiler stuff". By extension,
this makes the example actually work and fixes a build error with the
elaborator complaining about multiple implementations of a top-level
generator.

modular-orig-commit: f0b73867f76830456509e7ed4ff7b1b29c83c115