[mojo-stdlib][docs] Fix Matmul notebook and add moveinit to object (#15704)
The latest release broke the matmul notebook because `object` was given
actual lifetimes, meaning they have to be extended. The result was only
a `1.1x` improvement over python performance for `matmul_untyped`, which
is worse than the `8x` before. Adding a move constructor and changing
some of the method definitions to `fn` improved this back to `4x`, but
there is still room for improvement to get the `8x` back. This is
because Mojo is ineffective at removing unneeded copies.

Fixes [Internal Link]

modular-orig-commit: debff18f2ad90564d61e0d665defdac014312fd1