[Stdlib] Create a memory package to contain pointer, buffer, and memset (#19615)
This creates a new memory package which contains the unsafe pointers,
memset/memcpy, and the buffer/ndbuffer. We can think about placing
buffer/ndbuffer into their own package (called view), but let's put them
in memory for now.

This is also disabling the memset small, since packaging searchable
functions is not something we currently support and it is something that
we need to discuss.

modular-orig-commit: 34de8270be3fe5ca23035e76bc5be409f0312e6a