[mojo-sdk][build][ci] Add build target and workflow to publish examples (#19611)
When we release the Mojo SDK, we will release our notebooks and some new
code examples to the public "mojo" repo. This is the build command to
prepare those files and the github workflow to commit them with manual
dispatch only. However it's important that we run this workflow **from
the `mojo-docs-release` branch only**; **it should never be run from
`main`.** (I don't know of a way to enforce this other than to just
remove/disable the workflow in the main branch.)

Currently, this uses a new private repo called "mojo-test" just to
verify the github workflow works as expected. When we're ready for
public release, we'll update this to just "mojo".

And just to seed the new code examples directory, I added the "hello
mojo" example by @dualvtable ... More code examples are coming soon.

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>

modular-orig-commit: 9aa2ac5e87c198a8d539acbe52d9755a7c9f20b2