[docs] Add Mojo SDK docs to the main nav (#20473)
This included just about everything we need to release the Mojo SDK to
the public, including a revision to the "Hello Mojo" notebook to make is
suitable as a "language basics" guide that follows the "Hello world"
page.

It also combines the previous "Try Mojo" and "Install Mojo" docs for a
unified "Get started with Mojo" page, so as to boost the page rank and
avoid confusion with two separate pages to get started.

Plus other appropriate updates, such as move the SDK known issues to the
roadmap document and add the Mojo CLI docs to the nav.

**NOTE TO REVIEWERS:** I'm going to merge this without review in order
to stage it on dev-docs.modular.com and continue working on some other
incremental improvements. Please continue to send comments on these
files and I will integrate feedback with follow-up changes. Thank you!

re: [Internal Link]

modular-orig-commit: 3be6305a8baf6abf37ce108c88ed446aceaf16cd