[Stdlib] Properly free globals on program termination (#21649)
This makes sure that we free any globals that we allocate in Mojo upon
exit from main. Right now this is concentrating on the Python global,
but there will be other globals that will join this such as the random
number generator, global runtime, and any other singletons we want in
Mojo.

modular-orig-commit: 20f816f972b239f26f34505b1d6f68bc108c21f1