[docs] Clarify use of __bool__ in BoolMLIR notebook (#22625)
This was a bit misleading because it said that Bool does not implement
`__bool__`, but it actually does (perhaps just as a formality). I
decided it's easier and more clear to just avoid talking about what the
stdlib Bool does or doesn't do and focus on making OurBool not rely upon
it.

modular-orig-commit: d1de9a55e8d00c9b9f73e4cbdea82761904fbc2f