[mojo-lang, mojo-stdlib] Project var-arg into VariadicList (#23240)
A `VariadicList[T]` is automatically created for vararg `*arg: T` inside
the function body.

Majority of changes is cleaning up existing uses in mojo code. Places
that previously created VariadicList wrappers for varargs no longer need
the wrapper. Some functions have both var-arg and VariadicList overloads
that may no longer be needed, but that's still kept as is in this PR.

The PR only adds support for value arguments, not parameters.

modular-orig-commit: ac7fcec20ec1efcfa653f4a2da3aa3a655dcea2a