[Docs][Mojo] Remove bug note and reorganize docs for automatic function (#23839)
... parameterization.

Removed the separate section on "Accessing input parameters of a
parameterized type" and folded it into the automatic function
parameterization doc. It seems to me this is not a big standalone
feature, but a necessary prerequisite for automatic parameterization.
Please let me know if this should be left as its own section.

Note: Much easier to read with Jupyter notebooks rich diffs.

modular-orig-commit: 4f8b0bbea1c377ea40b2519fcbbf8d3d08b4c2a9