[docs] Consolidate Mojo docs into /docs/external/ (#25536)
The Mojo changelog has moved! All the `KGEN/docs/public/` files have
moved to `docs/external/mojo/`.

All doc build results should be the same; there should be no effect on
the docs website.

This is good for repo hygiene, to unify the existing docs and the new
Mojo manual. This is also helpful to enable the eventual open-sourcing
of the Mojo docs.

modular-orig-commit: 823a06a8b429646ddc0280d6dc5102b2bf87b62c