[docs] Rewrite the Mojo basics section (#25884)
This is the first page that follows the "Hello world" guide. It's
basically a new approach to the "Hello Mojo" notebook (which currently
takes this place after "Hello world"), but it's more concise because it
doesn't serve Mojo Playground users.

This version provides a terse intro to each topic that has its own page
in the Basics section (excluding packaging), and it also introduces
Python integration. The goal is to provide an quick overview of the key
features so users know their way around code, and then refer them to the
other sections of the manual for details.

Staged here:
https://solid-bassoon-wl2pkwr.pages.github.io/mojo/manual/basics/

I hope that this serves as a suitable replacement for the HelloMojo
notebook, which I think can be now deleted. I wrote that notebook as a
more gentle introduction to Mojo for the Mojo Playground, because the
Programming Manual was far too heavy. It also has worked well as a
language basics thus far, while the Mojo Manual is still in progress. So
I think HelloMojo is now obsolete and maintaining it in addition to the
Mojo Manual is not sustainable.

The bulk of this PR is about the "language basics" page, but it does
include some light copy edits to the "hello world" page, and a fix to
the CMake target because this filename conflicts with where we were
putting the HelloMojo notebook (so we shouldn't move that file over
anymore).

---

**TIP for notebook reviews:** Be sure you have [rich diffs
enabled](https://www.loom.com/share/bfa3e9ec28c44f509be0fcfa7ec0a245),
and then you can toggle between JSON source (required to make inline
comments) and the rendered "rich diff" mode:

<img width="417" alt="Screenshot 2023-10-02 at 8 42 03 AM"
[Internal Link]

modular-orig-commit: 42cd11b93b4fee2c91a0e23a94c4e2061d0a2cfb