[mojo-stdlib] Improve perf of string split function (#26606)
The naive implementation done in [Internal Link]
results in a lot of small string allocations due to constantly appending
to form new word chains.

Improve the performance by searching for the delimiter and forming the
matches using string slicing syntax to avoid the constant string joining. This
also makes it easy to support delimiters that are more than a single
character in this algorithm. So, remove that restriction and add a test case for when
the delimiter is multiple characters.

Update the mojo changelog to mention the new `split()` function now that
the performance isn't abysmal.

modular-orig-commit: c6130a12c6102fcfcad328b47096ec6accb4f191