[docs] Add CONTRIBUTING and CODE_OF_CONDUCT (#26287)
These copies are meant for internal version control and team review
purposes. All docs in this directory must also be committed to the
public repo, separately. Merging them here is merely a convenience so we
can review changes and discuss them in private before pushing them to
the public.

regarding: [Internal Link]

modular-orig-commit: 4d046b5e10292a7cc6e67ea9266ef9362535fa91