[Docs] More Traits info from changelog. Also small TOC reorg. (#26779)
Added several items from the changelog that weren't described in the
doc: static methods, handling register-passable types, and the list of
stdlib traits.

Long-term we won't keep the list of traits in this doc, but until our
API doc tools are more mature, we don't have any way to find traits in
the standard library.

I also noted that I was using "implements" and "conforms to"
interchangeably and it was somewhat confusing. I tried to make the
language more consistent. That is, a struct **conforms to** a trait. It
does so by **implementing** the requirements of that trait.

Not a huge change but staged so you can see it in context:


https://symmetrical-adventure-kq1pz3n.pages.github.io/mojo/manual/traits/

modular-orig-commit: b78f96a7501bd558e1f11cec1a5d97a7a7f979d9