[docs] Remove custom HTML styles for Docusaurus compat (#27246)
This is necessary to fix Docusaurus builds.

Unfortunately, we can't do custom HTML styles this way. It is possible
in Docusaurus, but it needs to be MDX style, not HTML. Thus, we can't do
this in a way that's compatible with BOTH Quarto and Docusaurus. It's
better if we just put the errors in comments for now and then figure out
a better style when we're fully migrated to Docusaurus.

modular-orig-commit: 078ed37140aac09c61fa4aa3b07d1826364f4154