[mojo-stdlib] Allow users to opt in to debug assertions (#27924)
As discussed in [Internal Link] right now,
external users have *no way* to enable assertions for safety sake in the
standard library.  They get assertions always from the mojo compiler which is
great, but nothing from the standard library side of things.  This results in a
poor user experience since the program just crashes with no information
typically.

Currently, a lot of the assertions in the standard library are defined in terms
of `debug_assert`.  So, hijack this a bit to give users a hook to opt-in for
assertions.  Assertions can be enabled when compiling a Mojo program by
specifying `-D MOJO_ENABLE_ASSERTIONS`.  This is good for both internal and
external users.  In the internal case, it doesn't mandate a debug kernels
build.  As a standard library author, I want a release build of kernels, but I
would love for assertions to be enabled still in the library.  For external
users, they have an opt-in ability to get standard library assertions when
compiling their Mojo source files.

Note:
- This is still off by default and purely opt-in.

Fixes [Internal Link]

modular-orig-commit: e7953421417ca96872e01fb9a45aa0bb84e6782b