[LSP] Add support for the references request (#27438)
This was easier than expected but we are limited to the current
document.
The main implementation detail is that we now keep a mapping of Symbol
to SymbolRefs. Other than that the rest of the implementation is
trivial.
I tried to add a python unit test but it hangs. I'll take a look at it
later.

modular-orig-commit: 4137febb2fcb8e2bfce9527e67e952d2fee59a6c