[mojo-stdlib] Default Runtime to use num_threads and add functions to llcl (#28566)
There were two performance issues for parallel code by default, without
a way for a Mojo user to rectify them. This PR defaults to max
performance, but still allows the user to configure to old behavior.

 ## Hyperthreading
The default implementation of the Global runtime was not using all
available threads, instead it was only using the physical cores, which
can have significant impacts on benchmarks. This has been rectified to
using all threads by default. For old behavior, a user can run:
 
 ```python
 with llcl.Runtime(llcl.num_cores()) as rt:
    parallelize[foo](rt, C.rows)
 ``` 
 
## Apple Performance Cores
On Apple machines it was using just the P (performance) cores by
default, reducing the total number of threads/workers utilized by the
`llcl` runtime. This was unchangeable for a Mojo user. This PR makes the
default to use all performance and efficiency cores, the user can then
easily oversubscribe the `parallelize` function to maximize performance.
If a user wants to only use performance cores, there is a new function
added to `llcl`:

```python
with llcl.Runtime(llcl.num_performance_cores()) as rt:
    parallelize[foo](rt, C.rows)
```

## Multiple ways of querying threads / cores
There were multiple functions and methods being used to determine the
number of threads and cores, this moves it all to just using
`llvm::harware_concurrency` which can take into account sockets and
hyperthreading and works across platforms.

Addresses [Internal Link]

modular-orig-commit: a9a7b3b7694654e10b3938676246206cc784e086