[mojo-stdlib][docs] Make cpython an internal module and fix missing documentation (#28641)
While changing documentation to be generated from packages,
I stumbled upon a bunch of modules not actually encoded in cmake,
which hid them from being properly checked for documentation. This
PR fixes most of that, but for the cpython module just makes it internal
(given it has no documentation at all, and isn't super user facing).

modular-orig-commit: fdb549dc014180cbff1e09cee104bf0ced7a4e9e