[mojo-lang][CheckLifetimes] Massive refactor of CheckLifetimes, ~~NFC (#28782)
The CheckLifetimes pass does two different walks over the IR: the
first is a forward bitvector dataflow passes that makes sure values
are defined before use, and the second is a backward dataflow pass
that decides where lifetimes end to insert destructors.

Both of these walks need to understand detailed information about
how the operations use and define and consume their operands, whether
they are indirect, etc etc etc.  Previously this was implemented by
each walk enumerating all of the operations it understands and "doing
stuff".  This might have made sense when we have 4-5 operations that
matter, but today we have dozens of them, with subtle interactions,
and they need to tightly interlock (or else we get difficult to debug
problems).

This patch changes the approach to factor out knowledge of the
"operand effects" of each operation into a ~~declarative specification
in the LifetimeTrackable.cpp file, and changes both walks over the IR
to use that single source of truth.  While doing this, it also adds a
ton of assertions, which has been catching all the weird and subtle
bugs that I've been fixing over the last couple days.

This design is better because it has SSOT, but also because it splits
out the "behavior of the ops" logic from the "mechanics of
checklifetimes".  In principle, we could even make the logic open and
extensible with an operation interface someday if there were
motivation to.

This is supposed to be an NFC patch, but as you can see there are a
couple of specific changes necessitated by this, which are
progressions in QoI and detecting a problem in object.mojo that went
undiagnosed before.

modular-orig-commit: a9461c6038412ab4f91a424923e74fb912eb4818