[mojo] Remove `@adaptive` and autotuning examples from public docs (#28777)
We advertised autotuning as part of the Mojo May launch as a core
feature of the language, but the reality is that it's underinvested and
underused even internally. It's certainly not something we actually want
external users to be using, and from external bug reports, it doesn't
seem to work most of the time.

Remove the public examples and indicate in the changelog that it's
undergoing a redsign but not being removed from the language.

modular-orig-commit: 689a9aeac90af75d9391f15deaab8ba6066345f1