[docs] remove packages from the modules index (#28886)
Currently, the Module Index page (`lib.md`) lists all files in `stdlib`,
so it also shows the new package summary pages. This changes that so
that the glob search ignores `index.md` files.

Also attempt to print "package" for the subtitle, but that's not quite
working as-is. I think it's a bug somewhere upstream, so I think the
change in this commit _should_ work. See:
[Internal Link]
UPDATE: Found an alternative solution in
[Internal Link]

modular-orig-commit: 0a0e1cbbb462fa40c1acf09dec44a6784203e7e6