[mojo-lang] Implement support for first-level indirect references. (#29008)
Well this is a big thing - we can now reason about indirect references
in destructor elimination, which means that we properly extend the
lifetimes of things indirectly referenced, and we handle overwrites
of mutable values through references correctly.  This is (finally)
the core feature set that makes references /interesting/ after weeks
of yak shaving.  The testcases show this working.

There is one major caveat, and one minor thing:
1) This only handles direct uses of low-level `!lit.ref` types, not
   things that wrap them up in e.g. a `Reference` struct type.  This
   is a critical next step.
2) We are /almost/ correct but not completely correct with lifetime
   tracking, e.g. the `__result__` slot of throwing functions is
   wrong.  This won't matter in practice given the grotty workaround
   I added, but I'd like to eventually remove this.

This unrelatedly deletes a dead method from coroutine.

modular-orig-commit: 41ace3f1670cb44c4a0b7acb7aac42f6926fa0fb