[mojo] `mojo package` never pre-elaborates (#28991)
In #28631, `mojo package` was modified to no longer include package
archives by default, and only include them when certain "targeting"
options were used. But this behavior is vague and unintuitive; #28774
(not landed) highlights how difficult this is to explain in
documentation, and #28820 demonstrates how easy it is to accidentally
use one of these options and end up including a package archive.

My original intention was to leave this logic in because I wanted it to
be exercised, but this was overly sentimental -- in fact, full
elaboration and compilation occurs when the package is imported and
compiled on-demand. There's no need to keep the package archive
pre-compilation logic around just to keep it working with regression
tests. This patch removes it completely. It also removes target options
from the set of options `mojo package` accepts, since they would no
longer do anything.

Closes #28027.

modular-orig-commit: 887937f3dadab8a806ae4256f2e8b8522bed1135