[mojo-lang] Pull back `__get_ref_from_value` significantly. (#29068)
This removes all of the uses of `__get_ref_from_value` from the standard
library, and removes the changelog documentation.  This is now only used
by the CheckLifetimes related testcases that need to exercise very
specific low-level things.  I considered removing it entirely, but these
tests are too important to lose.

As part of this, this moves `variant.mojo` off of this, using the
`Reference` type instead.  As part of that, I added a
`get_unsafe_pointer()` method to `Reference`.  This helper would ideally
be an init on `Pointer` itself, but we can't do that yet because of
`AnyRefType` vs `AnyType` differences.

modular-orig-commit: 6e97883d3d02f4a479b67f5e36ba3efd7e083cdb