[mojo-stdlib] `collections.dict.Dict` (#27811)
An efficient, O(1) amortized average-time complexity for insert, lookup,
    and removal of dictionary elements.

    Usage:

    ```mojo
    var d = Dict[String, Int]()
    d["a"] = 1
    d["b"] = 2
    print(len(d))      # prints 2
    print(d["a"])      # prints 1
    print(d.pop("b"))  # prints 2
    print(len(d))      # prints 1
    ```

    Its implementation closely mirrors Python's `dict` implementation:

- Performance and size are heavily optimized for small dictionaries, but
can
        scale to large dictionaries.
- Insertion order is implicitly preserved. Once `__iter__` is
implemented
        it will return a deterministic order based on insertion.
- To achieve this, elements are stored in a dense array. Inserting a new
element will append it to the entry list, and then that index will be
stored
in the dict's index hash map. Removing an element updates that index to
a special `REMOVED` value for correctness of the probing sequence, and
the entry in the entry list is marked as removed and the revelant data
is freed.
The entry cannot be re-used until the dictionary is either resized or
compacted.
- The index probe sequence is taken directly from Python's dict
implementation:
      ```mojo
      var slot = hash(key) % self.reserved
      var perturb = hash(key)
      while True:
        check_slot(slot)
        alias PERTURB_SHIFT = 5
        perterb >>= PERTURB_SHIFT
        slot = ((5 * slot) + perturb + 1) % self.reserved
      ```
- Similarly to Python, we aim for a maximum load of 2/3, after which we
resize
        to a larger dictionary.
- In the case where many entries are being added and removed, the
dictionary
can fill up with `REMOVED` entries without being resized. In this case
we will eventually "compact" the dictionary and shift entries towards
        the beginning to free new space while retaining insertion order.

Key elements must implement the `KeyElement` trait, which encompasses
Movable, Hashable, and EqualsComparable. It also includes
CollectionElement
    and Copyable until we have references.

Value elements must be CollectionElements for a similar reason. Both key
and
value types must always be Movable so we can resize the dictionary as it
grows.

Without conditional trait conformance, making a `__str__` representation
for
Dict is tricky. We'd need to add `Stringable` to the requirements for
keys
    and values. This may be worth it.

modular-orig-commit: 735f338d1e98c607c0af1a6da032aa10cbe45108