[Stdlib] Fix floating-point SIMD to int conversion (#29269)
Floating point to int conversion in python does not round towards zero
(floor). Instead it's just a C -cast operation and discards the
fractional part of the value. This means that any decimal part of the
float will be discarded.

For example, converting the float 390.8 to an int will result in the
integer 390. The decimal part, .8, will be discarded. Another important
thing to keep in mind is that if the float is negative, the resulting
int will also be negative. For example, converting the float -390.8 to
an int will result in the int -390.

modular-orig-commit: a5601476d387635bc60bf664e38f7e5d7ebcf97e