[mojo-lang] Sort out the closure model (#29385)
This PR takes the penultimate step to rolling out the new runtime
closures across the language and works towards having a cleaner model
that balances parameter closures and runtime closures.

This PR removes the `@closure` and `@noncapturing` decorators. When
declaring a nested function, if the function does not capture any
dynamic values, it is a bare function by default. If it captures any
dynamic value, it becomes a closure. The new closure is memory-safe and
contains a heap-allocated capture list. The default capture semantics is
capture-by-copy. For internal users, you can use the
`@__move_capture(x, y)` decorator to get move capture semantics until
capture list syntax is a thing. The type of a runtime closure is
`fn(...) escaping -> T`. (keyword still WIP)

To get a `capturing` parameter closure, use the `@parameter` decorator.
This will make any function, nested or top-level, `capturing`.

modular-orig-commit: 38c51f40e018f65059a11e1b933b95e70c14626c