[mojo-lang] Teach CheckLifetimes about lifetimes. (#29508)
This (finally) teaches CheckLifetimes about indirect accesses through
lifetime bound types, using the mutability specifier on the bound
lifetimes to track what sort of access is happening.  This allows us
to extend the lifetime in the testcase, fixing a dangling pointer
problem and making `owned` mem variadics fully correct (AFAIK).

There is still an issue with dtor ordering, but it doesn't affect
the variadics case.

This PR also cleans up the changelog description of Reference.

modular-orig-commit: 49d2b71afcf1648b482bca517738195f6416d41f