[Mojo] Add Copy Capture Decorator to Enable Capture By Copy for Vars (#28487)
Introduce a `__copy_capture` decorator to allow capture of vars by
immutable copy. This is intended to enable let captures in async
parametric functions to migrate to var captures for the purposes of
removing let syntax. It can be used with escaping closures as well but
since capture by copy is the default there is no reason to use this with
escaping closures. It just shouldn't break.

modular-orig-commit: 61332c89de57714ac48387bb5b220def710402c8