[mojo-stdlib] Compartmentalize stdlib tests (Renamed) (#30027)
(This was #29084, but that was failing CI because the `#` in the branch
name wasn't valid as a docker tag.)

This aims to resolve [Internal Link]
i.e. consolidating all of the Mojo standard library tests into their own
subdirectory, inside Kernels/test/stdlib. This is a further step to
distinguishing the stdlib from Kernels more generally.

modular-orig-commit: 2f5db6024ee7e8c838c7376884d19d872efb395b