[mojo-lang] Fix problem in LowerLIT, adopt `__lifetimeof(x)` (#30133)
This fixes LowerLIT to know that return ops potentially need their
operands substituted when implicit lifetimes are eliminated, because
they can return a reference or other type that includes one!

This change allows us to use `__lifetimeof(self)` to dissolve one
crime I introduced in the path to get variadics to work.  Another
remains but the type checker isn't ready for it.

modular-orig-commit: df0a6e696973f5e0b4838b0a0a35cdec9999e0dc