[mojo-lang] Refactor fn signature parsing to eliminate `ParamConstantOp`. (#30138)
Function signature parsing needs to create IR objects so that arguments
that are dependent on each other can succeed at lookup and have
reasonable
IRValues that indicate their value rep and type.  We previously were
using temporarily allocated `ParamConstantOp` to do this work, which
were
then deallocated after the signature was parsed.  For function bodies,
we'd then turn around and allocate a bunch of BlockArguments.

Switch to creating the block arguments directly into the function during
type checking. This eliminates redundant work and a bunch of
allocations.
This also encourages us to push things to be more explicit in the
argument
list, e.g. a rework of the variadic passing conventions.

Also document `__lifetime_of` in the changelog.

modular-orig-commit: 5d7da627873717cc4f1d8322429ef5ab7e5e014e