[mojo-stdlib] Rename `lldb.mojo` to `visualizers.mojo` (#31428)
As discussed in [Internal Link]
rename `lldb.mojo` to `visualizers.mojo` because the REPL also uses
these visualizers and it's not obvious to external users that the REPL uses
LLDB under the hood. This is just a little clean up as we move toward making the
`debug` module public in prep for open sourcing the stdlib.

modular-orig-commit: a089fd3466dcc14d88d3c009b398e1e6c9c758b4