[mojo-stdlib] Document reason for explicit print overloads (#31528)
Now that `print` calls a value's `__str__` function, and a lot more
types in the stdlib declare conformance to the `Stringable` trait, we *could* in
theory remove a lot of the explicit overloads. However, they're defined
explicitly to reduce binary size and for GPU codepaths where the variadic print
doesn't work. So, document this rationale to avoid future developers falling in this
pitfall and trying to remove them.

modular-orig-commit: ee5848338f92a3d539e99e944ceb418f51334509