[mojo-stdlib] Make Arc reg-passable (#31718)
`Arc` just holds a pointer, so it can be made reg-passable.  So, mark it as
such.  This will be useful as we may rework some of the tensor lifetime woes in
the engine Mojo API to work with `Arc[Tensor]`. Unfortunately, keeping the
`@value` decorator even when defining our own `copyinit` causes crashes, which
smells like a compiler bug.  So, remove the `@value` decorator for now.

modular-orig-commit: e894483d46f8e1489da82b378ca98b06b87d03c1