[mojo-stdlib] Allow specifying `end` for `print()` (#32542)
For parity with Python and hopes of being able to remove `print_no_newline`,
add a keyword-only `end` argument to `print` which is printed after printing the
elements.  By default, as before, print the newline character.  This gives
flexibility to the callers, and we can transition all of the existing callers of
`print_no_newline` over to `print(..., end = '')` for example.

In the future, we may want to provide a `sep` keyword-only argument for parity
with Python.  This is easy to implement, and I've marked a `TODO` for where we
need to thread things through to have callers specify the separator between
printing elements.

Note:
- We can't just modify `print_no_newline` to do:

```
print(first, *rest, end = "")
```

since this results in an expected error:

```
error: unpacked arguments are not supported yet
    print(first, *rest, end = "")
```

Related to [Internal Link]

modular-orig-commit: b612eb1ff37459909813cf83c7c3496dfb05f1b4