[mojo-stdlib] Move `MoveOnly` to `test_utils` package (#32522)
As discussed in [Internal Link] it's clear we
need to ship the `test_utils` package alongside the `stdlib` since other
`stdlib` tests rely on some grab-bag of functions in this package.

So, for convenience, move the `MoveOnly` type to this `test_utils` package too.
This means the things in the `testing` module in the `stdlib` package are the
"unit testing framework" parts only and then the misc testing
functions/utilities are in the `test_utils` package.

Fixes [Internal Link]

modular-orig-commit: be1cc3807a8aa0c066be72f941db8f971d0a55ca