[mojo-lsp] Add support for generating doc string templates (#32593)
This makes it a lot easier to write documentation strings, as
it will generate the expected template when a decl has an
empty doc string. Right now we just generate the template,
but when lsp 3.18 lands, we can turn the template into a
snippet which will enable better tab completion when filling
it out.

Closes #25151

modular-orig-commit: 9670a7c57c2a7c7a790b450c019d670fd10a5cd3