[Docs] Add *_ expression, update automatic parameterization docs. (#32612)
Moves the section on "Fully-bound, partially-bound, and unbound types"
later in the doc, between alias and automatic parameterization where it
actually makes a little more sense. And updated the automatic
parameterization section to focus on explitly unbound parameters, and
moving the older, implicit unbound parameter syntax to a subsection.

Also removed lets from this document. This requires a bit of care so I'm
doing it lazily as I update files.

Staged:
[Internal Link]

modular-orig-commit: b41f515b9948517cf171788256313e3ea91921a4