[mojo-lsp] Turn off `mainWillDonate` for the lsp runtime (#32725)
Main donation was accidentally turned on in #28280, but we
never want to block the main thread, and never call LLCL::await.
This effectively has meant that a lot of LSP requests were added
to the main thread to run, but never actually run. This breaks a
lot of requests seemingly randomly, and can make the LSP really
flaky.

Fixes [Internal Link]

modular-orig-commit: f342683c8964c7b0722052a523bc1bf806bc977f