[mojo-lang] Remove the `-mojo-experimental-lifetimes` flag + support. (#32867)
This flag enabled syntactic sugar for references, and applied that
sugar to the `!lit.ref` type.  We now have the `Reference` type,
and with other general language features we think we can probably
get away without reference-specific language sugar/support.  This
means we can give the identifiers back to the user instead of
reserving them as keywords.

modular-orig-commit: 5faef3811113fdf2782cfa6a57d24e952d971d4c