[mojo-lang] rename FloatLiteral to FloatLiteralOld (#30639)
as part of the infinite precision FloatLiteral work described here:
[Internal Link]

We need to rename FloatLiteral to FloatLiteralOld because the new FloatLiteral needs a runtime type to materialize to.  Without making various changes to the SIMD struct, we can't have the new FloatLiteral target Float64 directly without many conversion errors.  We will eventually remove this FloatLiteralOld type in favor of SIMD's Float64 type, but this is an incremental step.

modular-orig-commit: 122771aed2324b34a66a673f3653bcced0684137