[mojo-stdlib] Add initial lit config file for OSS (#32936)
This adds an initial starting point for our lit config file uses in the
open source repo for running the tests. Since we already have an
internal `lit.cfg.py`, choose a different name for the slight
differences between internal and external configurations.

The `lit` tool provides a way to specify name for the config file
(outside of the default of `lit.cfg.py` for example). To use the new
config:

```
lit --config-prefix=open-source-lit <other_lit_options>
```

for example.

Related to [Internal Link]

modular-orig-commit: 4391429b2aef6a85f478a8f867ce8d45bacb9308