[mojo-lang] remove FloatLiteralOld (#32939)
This PR removes the `FloatLiteralOld` type that was introduced when moving to the new infinite precision `FloatLiteral` type, making `FloatLiteral` materialize to `Float64`, the SIMD scalar type.

At present this PR disables a test because Float64's (IE SIMD's)`__pow__` method doesn't work with a negative LHS and a fractional RHS, returning NAN whenit should be a particular negative normal value.  This is a bug in SIMD, and isn't important enough to block removal of `FloatLiteralOld`.  The problem is tracked here:  [Internal Link]


modular-orig-commit: 25d994c11a510799577aff0408da11d50b5948ec