[mojo-stdlib] Add `apple-silicon` feature in sys lit config (#33916)
`sys/test_aarch64_target.mojo` has a `REQUIRES:` directive on
`apple-silicon`, but that feature is only set in
`Kernels/test/linalg/lit.local.cfg`. The `apple-silicon` feature is used
by a linalg test. For now, just copy the few lines of code in the lit
config to add the necessary `apple-sillicon` feature so this test runs
now in `sys`.

modular-orig-commit: 420fc9fc961ad4372564e6d3291a2decca846a82