[Mojo] Do not allow type `PValue` materialization (#33897)
Recently, we decided that supporting type variables and type functions
is posing a large surface area for bugs and crashes, and since type
functions are not actually very useful (due to lots of missing features,
e.g. existential) it's better to just disable this for now. This patch
does that and also attempts to add some suggestions for fixes in the
error messages.

modular-orig-commit: 0d0a6521f1a3fefa87dadd72d0f8ec8e299c3d78