[mojo-stdlib] Remove use of `Tensor` in `test_file.mojo` (#34352)
`read_bytes` returns a `List` now, so there's no need to round-trip
through `Tensor` still. So, just construct a `String` out of the `List`
now instead to remove the dependency on `Tensor`.

Note:
- Can't use `StringRef` as-is since there's no constructor from an
`AnyPointer` which is the backing data for `List`. We should add that
constructor in a follow-up.

Fixed [Internal Link]

modular-orig-commit: 69d30c4e269b75c84936221aeadbb68aa05fc48c