[CI] Ensure `mojo format` in lint/format steps
Add a step in the build and test workflow for ensuring any added or modified
Mojo code adheres to `mojo format`.  Fail the step (and the GH workflow) if any
code does not adhere to our existing formatting rules.  Right now, there's no
`pyproject.toml` that `mojo format` or `mblack` uses under the hood — it's just
the default line length of `80` set by `mojo format` and then the normal
defaults of `mblack`.

The step is set up so we can easily extend it to running `markdownlint` for
Markdown files in the future.