[mojo-stdlib] Shuffle around stdlib tests that depend on non-OSS modules (#35487)
The process of splitting stdlib modules into those we'll be including in
the initial OSS release and those we wont left many tests behind in
Kernels/test/stdlib that came to reference non-OSS modules.

We want the OSS tests to reference only OSS stdlib modules. So this PR
shuffles around individual test cases in the OSS tests that use non-OSS
stdlib functionality, moving them to the non-OSS test suite.

Fixes #34925.

modular-orig-commit: 5a8f089f0b18152ed72f31d37054a8a70f87759a