[mojo-stdlib] Check for `not` binary in `lit.cfg.py` (#35633)
Not all distros ship the `not` binary with the LLVM installation.  For example,
macOS with `brew install llvm` does, but LLVM from `apt` on Linux does not.
Ditto with the download from `https://apt.llvm.org/llvm.sh`.  Fun.

So, add a check for detecting whether the `not` binary is available (and in the
`PATH`).  If so, add a feature: `has_not`.  Ensure the relevant tests require
this new `has_not` feature so these tests get marked as `UNSUPPORTED` for those
trying to run the tests but do not have the `not` binary.  The previous behavior
is that the test would fail.  Now, it is just skipped.

modular-orig-commit: e9ff5f7244eac0b654975cda6d7aceab9822d5f9