[Utils] Teach `lint-copyright-notices.py` about public license notices (#35682)
Files with public licenses have been added to the ignore list of the
license checker. This obviously lead to some of these files having
incorrect or missing license notices. This patch fixes this script so
that it now has a list of files it should treat as public. All changes
except the script itself are NFC, but please review to ensure they now
got the correct license.

Closes [Internal Link]

modular-orig-commit: 9018362dca667d98efd8db69bfd3273c92c39ac4